+++ src/main/java/org/example/insurance_card_be/controller/dashboard/SaveInsuranceCardInfoServlet.java
@@ -1,0 +1,20 @@
+package org.example.insurance_card_be.controller.dashboard;
+
+import jakarta.servlet.ServletException;
+import jakarta.servlet.annotation.WebServlet;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+import java.io.IOException;
+
+@WebServlet("/saveInsuranceCardInfo")
+public class SaveInsuranceCardInfoServlet extends HttpServlet {
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        String cardNumber = request.getParameter("cardNumber");
+        String issueDate = request.getParameter("issueDate");
+        String expiryDate = request.getParameter("expiryDate");
+
+
+    }
+}
